zebra_project/requirements.txt
certifi==2022.6.15
cycler==0.11.0
efficientnet-pytorch==0.7.1
fonttools==4.33.3
joblib==1.1.0
kiwisolver==1.4.3
matplotlib==3.5.2
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
numpy==1.22.3
# numpy-base==1.21.2
olefile==0.46
opencv-python==4.6.0.66
packaging==21.3
pandas==1.4.3  
Pillow==7.1.2
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
threadpoolctl==3.1.0
# torch==1.8.0
# # torchaudio==0.8.0a0+a751e1d
# torchaudio==0.8.0
# torchvision==0.9.0
tqdm==4.64.0
typing_extensions==3.10.0.2
